+++ server/routes/html.js
@@ -20,7 +20,7 @@ export default () => {
       <link rel="stylesheet" type="text/css" href="https://cdn.auth0.com/manage/v0.3.1672/css/index.min.css" />
       <link rel="stylesheet" type="text/css" href="https://cdn.auth0.com/styleguide/4.6.13/index.min.css" />
       <% if (assets.style) { %><link rel="stylesheet" type="text/css" href="/app/<%= assets.style %>" /><% } %>
-      <% if (assets.version) { %><link rel="stylesheet" type="text/css" href="<%= assets.cdnPath %>/auth0-delegated-admin.ui.<%= assets.version %>.css" /><% } %>
+      <% if (assets.version) { %><link rel="stylesheet" type="text/css" href="<%= assets.cdnPath %>/auth0-criipto-verify-admin.ui.<%= assets.version %>.css" /><% } %>
       <% if (assets.customCss) { %><link rel="stylesheet" type="text/css" href="<%= assets.customCss %>" /><% } %>
     </head>
     <body>
@@ -31,8 +31,8 @@ export default () => {
       <% if (assets.vendors) { %><script type="text/javascript" src="/app/<%= assets.vendors %>"></script><% } %>
       <% if (assets.app) { %><script type="text/javascript" src="/app/<%= assets.app %>"></script><% } %>
       <% if (assets.version) { %>
-      <script type="text/javascript" src="<%= assets.cdnPath %>/auth0-delegated-admin.ui.vendors.<%= assets.version %>.js"></script>
-      <script type="text/javascript" src="<%= assets.cdnPath %>/auth0-delegated-admin.ui.<%= assets.version %>.js"></script>
+      <script type="text/javascript" src="<%= assets.cdnPath %>/auth0-criipto-verify-admin.ui.vendors.<%= assets.version %>.js"></script>
+      <script type="text/javascript" src="<%= assets.cdnPath %>/auth0-criipto-verify-admin.ui.<%= assets.version %>.js"></script>
       <% } %>
     </body>
     </html>
@@ -57,7 +57,7 @@ export default () => {
     // Render from CDN.
     const clientVersion = process.env.CLIENT_VERSION;
     if (clientVersion) {
-      const cdnPath = config('CDN_PATH') || '//cdn.auth0.com/extensions/auth0-delegated-admin/assets';
+      const cdnPath = config('CDN_PATH') || '//cdn.auth0.com/extensions/auth0-criipto-verify-admin/assets';
       return res.send(ejs.render(template, {
         config: settings,
         assets: {