+++ CHANGELOG.md
@@ -92,7 +92,7 @@
 
 ### Changes
 
-- Extension no longer uses to hash as a name (auth0-delegated-admin is used instead).
+- Extension no longer uses to hash as a name (auth0-criipto-verify-admin is used instead).
 
 ## [2.1.0] - 2016-10-10